Add tile caching and limit range

+++ pkg/coordinates/convert.go
@@ -12,8 +12,8 @@ func Degrees(rad float64) float64 {
 	return rad * 180 / math.Pi
 }
 
-func Y2lat(ty float64, zoom int8) float64 {
-	y := ty / math.Pow(2, float64(zoom))
+func Y2lat(ty int, zoom int) float64 {
+	y := float64(ty) / math.Pow(2, float64(zoom))
 	l := (1 - 2*y) * math.Pi
 	o := math.Atan(math.Sinh(l))
 	return Degrees(o)
@@ -23,8 +23,8 @@ func Lat2y(lat float64) float64 {
 	return R * math.Log(math.Tan(math.Pi/4+Radians(lat)/2))
 }
 
-func X2lon(tx float64, zoom int8) float64 {
-	x := tx / math.Pow(2, float64(zoom))
+func X2lon(tx int, zoom int) float64 {
+	x := float64(tx) / math.Pow(2, float64(zoom))
 	l := (2*x - 1) * math.Pi
 	return Degrees(l)
 }
